# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The REST API for GitHub Gist Backend",
        "title": "GitHub Gist Backend REST API",
        "contact": {},
        "version": "1.0-alpha"
    },
    "basePath": "/api/",
    "paths": {
...
        "/gists/{gistId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gist Operations"
                ],
                "summary": "Get the gist by gist id, does not load gist comments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The ID of the gist",
                        "name": "gistId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/gists/{gistId}/comments": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gist Operations"
                ],
                "summary": "Get the comments of a gist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The ID of the gist",
                        "name": "gistId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}